This does away with the SDL mutexes, locks, and position coordination in favour of using back-pressure from the RW queue to pace the producer and consumer threads:
 - the mixer-callback thread waits until a rendered buffer is available before being fed samples
 - the rendering thread waits until space is available in the queue before rendering another buffer-full

Rendering is now performed in fixed-sized 4 KiB chunks, which is more efficient than rendering to smaller and/or per-millisecond chunks. The RW queue also performs `std::move()` operations on the rendered buffers to ensure no copies are created.

@dreamer , unfortunately I was unable to spread these changes over multiple commits; it was simply "one or the other" (without destroying compile-ability).  On the review side, I'm happy to annotate the changes in GitHub to help the review effort. 

I went ahead with this after the other PR involving benchmarking the Moody RW queue proved it to 8+ fold less-latent than a rotely implemented blocking queue. Regardless which queue we use, this design change significantly improves the code, which is where I see the most value. 

A welcome side-effect was fixing #858, which was caused by an SDL poll-wait loop experienced on some Aarch64 SBCs. Thanks to @SupervisedThinking for both reporting and testing.