@kcgen 2021-02-05
x86-64 CPU-limited to 800 Mhz:

Space Quest IV configured for MT-32 music, 9000 cycles, dynamic core, and 48 KHz mixer rate. 

Release build of dosbox-staging `master` verus this branch.

Measured using `time`, which tallies all sub-threads into the `user` and `sys` durations. Launched with:
  `time /usr/src/dosbox-staging/src/dosbox sq4.bat & sleep 60 && killall dosbox`

The first five runs testing the two binaries were recorded.

**`master` scores (sorted low-to-high)**
| user | sys | duration | (user + sys) / duration |
| -- | -- | -- | -- |
| 22.631 | 1.870 | 60.155 | 40.729 |
| 22.717 | 1.923 | 60.153 | 40.962 |
| 22.934 | 1.843 | 60.181 | 41.171 |
| 22.881 | 1.976 | 60.152 | 41.324 |
| 23.339 | 1.868 | 60.231 | 41.851 |

**PR scores (sorted low-to-high)**

| user | sys | duration | (user + sys) / duration |
| -- | -- | -- | -- |
| 22.721 | 1.712 | 60.153 | 40.619 |
| 22.677 | 1.939 | 60.197 | 40.892 |
| 22.979 | 1.787 | 60.182 | 41.152 |
| 22.935 | 1.848 | 60.203 | 41.166 |
| 23.225 |  1.814 | 60.187 | 41.602 |

Overall, the numbers are a wash and at least show no performance change or regression is caused by this PR.