@kcgen 2021-02-05
Raspberry Pi 3. Same build and test scenario as above.

**`master` scores (sorted low-to-high)**
| user(s) | sys(s) | wall(s) | (user + sys) / wall |
| -- | -- | -- | -- |
| 60.409 | 67.263 | 2.831 | 116.032379281233 |
| 60.322 | 67.831 | 2.661 | 116.859520572925 |
| 60.288 | 68.467 | 2.265 | 117.323513800425 |
| 60.238 | 69.289 | 2.698 | 119.504299611541 |
| 60.274 | 70.352 | 1.752 | 119.627036533165 |


**PR scores (sorted low-to-high)**

| user(s) | sys(s) | wall(s) | (user + sys) / wall |
| -- | -- | -- | -- |
| 60.316 | 68.574 | 1.673 | 116.464951256715 |
| 60.269 | 68.137 | 2.22 | 116.738289999834 |
| 60.262 | 68.628 | 1.872 | 116.989147389732 |
| 60.299 | 68.345 | 2.289 | 117.139587721189 |
| 60.32 | 69.733 | 2.157 | 119.181034482759 |

These numbers are also wash, and further the proof that  no performance change or regression is caused by this PR.