@kcgen 2021-02-05
This PR fixes #858, the SDL MT-32 SDL poll-wait loop experienced on the PINE64 RockPro64 and Amlogic A311D boards (both of which run Aarch64 kernels and 32-bit user-spaces).