@kcgen 2021-02-07
> can you verify if new implementation works well in this game? To run MT-32 sound effects paku.exe /mt32.

[paku_000.flac.zip](https://github.com/dosbox-staging/dosbox-staging/files/5939770/paku_000.flac.zip)

``` text
dosbox-staging version git
---
CONFIG: Loading primary file dosbox.conf
OPENGL: Pixel buffer object extension: available
SDL: Mouse will be captured immediately on start and middle-click will uncapture the mouse.
MEMORY: Base address: 0x7f6209666010
MEMORY: Using 256 DOS memory pages (1 MiB)
MIXER: Negotiated 2-channel 48000-Hz audio in 1024-frame blocks
MT32: Found ROM pair in ../roms/
MT32: Initialized CM-32L/LAPC-I Control v1.02 with CM-32L/CM-64/LAPC-I PCM ROM
MIXER: MT32 channel operating at 48000 Hz without resampling
MIDI: Opened device: mt32
MAPPER: Found no joysticks
MAPPER: Loaded default key bindings
MOUNT: Path '.' found
SDL: captured the mouse
MAIN: Draw resolution: 640x400, pixel aspect ratio: 1.20
MAIN: Pixel-perfect scaling (1x1): 640x400 (PAR 1.20) -> 640x400 (PAR 1.00)
MT32: LCD-Message:      PAKU  PAKU     
MT32: LCD-Message: ***  PAKU  PAKU  ***
MT32: LCD-Message:      PAKU  PAKU     
MT32: LCD-Message: ***  PAKU  PAKU  ***
MT32: LCD-Message:      PAKU  PAKU     
MT32: LCD-Message: ***  PAKU  PAKU  ***
MT32: LCD-Message:      PAKU  PAKU     
MT32: LCD-Message: ***  PAKU  PAKU  ***
MT32: LCD-Message:      PAKU  PAKU     
MT32: LCD-Message: ***  PAKU  PAKU  ***
MT32: LCD-Message:      PAKU  PAKU     
MT32: LCD-Message: ***  PAKU  PAKU  ***
```