@kcgen 2021-02-07
Squashed and inbound; thanks for the thorough review @dreamer .

> About PakuPaku - but what's the result of the test?

I visually can't see any latency differences between all three (MT32-master, MT32-RW-queue, MIDI).  

The dealth animation and symbol crash all appear when the little "X" vanishes with the symbols trailing.

[paku-side-by-side.zip](https://github.com/dosbox-staging/dosbox-staging/files/5940311/paku-side-by-side.zip)

If we want to, we can ratchet latency even tighter by shortening the buffer frames (from 2048 to to say 1024 or even 512), but then we'd only be rendering one or two frames in advance, which is a bit too close for comfort (atleast on systems like the Pi).